+++ CHANGELOG.md
@@ -30,7 +30,7 @@ at anytime.
   *
 
 ### Added
-  *
+  * is_address RPC method (checks if address is valid using lbryum)
   *
 
 ### Removed